Only show comments if the commenter shares a group with the current user
