+++ happiness-backend/api/token.py
@@ -1,45 +1,15 @@
 from apifairy import authenticate, response, other_responses
 from flask import Blueprint, request
-from flask_httpauth import HTTPBasicAuth, HTTPTokenAuth
 
+from api.auth import basic_auth, token_auth
 from api.models import Token
-from api.dao.users_dao import get_user_by_username, get_user_by_id, get_token
+from api.dao.users_dao import get_token
 
 from api.app import db
-from api.errors import error_response, failure_response
+from api.errors import failure_response
 from api.schema import TokenSchema
 
 token = Blueprint('token', __name__)
-
-basic_auth = HTTPBasicAuth()
-token_auth = HTTPTokenAuth()
-
-
-@basic_auth.verify_password
-def verify_password(username, password):
-    if username and password:
-        user = get_user_by_username(username)
-        if user and user.verify_password(password):
-            return user
-
-
-@basic_auth.error_handler
-def basic_auth_error(status):
-    return error_response(status)
-
-
-@token_auth.verify_token
-def verify_token(session_token):
-    if session_token:
-        token = get_token(session_token)
-        if token and token.verify():
-            return get_user_by_id(token.user_id)
-
-
-@token_auth.error_handler
-def token_auth_error(status):
-    return error_response(status)
-
 
 @token.post('/')
 @authenticate(basic_auth)
